requirements.txt
python-telegram-bot==13.7
psycopg2-binary==2.9.10
pytest==7.1.2
python-dotenv==1.0.0
matplotlib~=3.10.1
requests~=2.32.3

playsound~=1.3.0